+++ CHANGELOG.md
@@ -4,7 +4,18 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.9.18 - Unreleased
+## Version 0.9.19 - Unreleased
+
+### Changed
+* `grab_test_image` now falls back to mirrors if the first URL returns an error
+* Allow `format` to be positional in `kwimage.Box.coerce` and `kwimage.Boxes.coerce`
+
+### Fixed
+* Fixed compatibility with new scikit-image
+* Mixin methods are now shown in the docs
+
+
+## Version 0.9.18 - Released 2023-05-02
 
 ### Changed
 Disable torchvision NMS on windows by default